Fix 2/3 https://github.com/Euphillya/SkyFolia/issues/9 - class java.lang.Integer cannot be cast to class java.lang.Double in ConfigToml - Size exceeded ! Better documentation + fix island size limit

+++ api/src/main/java/fr/euphyllia/skyfolia/api/skyblock/Island.java
@@ -21,7 +21,7 @@ public abstract class Island {
 
     public abstract double getSize();
 
-    public abstract void setSize(int rayon) throws MaxIslandSizeExceedException;
+    public abstract void setSize(double rayon) throws MaxIslandSizeExceedException;
 
     public abstract @Nullable CopyOnWriteArrayList<WarpIsland> getWarps();
